+++ rdr_service/ppsc/ppsc_data_sync.py
